+++ userspace/lib/include/fcntl.h
@@ -1,6 +1,8 @@
 #ifndef SRVROS_POSIX_FCNTL_H
 #define SRVROS_POSIX_FCNTL_H
 
+#include <sys/types.h>
+
 #define O_RDONLY 0x0000
 #define O_WRONLY 0x0001
 #define O_RDWR 0x0002
@@ -18,6 +20,21 @@
 #define F_SETFD 2
 #define F_GETFL 3
 #define F_SETFL 4
+#define F_GETLK 5
+#define F_SETLK 6
+#define F_SETLKW 7
+
+#define F_RDLCK 1
+#define F_WRLCK 2
+#define F_UNLCK 3
+
+struct flock {
+    short l_type;
+    short l_whence;
+    off_t l_start;
+    off_t l_len;
+    pid_t l_pid;
+};
 
 int fcntl(int fd, int command, ...);